miner.py: implement checks and error handling for failing axons
The miner template seemingly assumes that starting an axon never fails.
In e2e tests the axon failed to start, due to mixing next_asyncio in
bittensor and regular asyncio in uvicorn. It would be better to
terminate the miner process when the axon never starts at all.

This patch addresses this by:
- wrapping run() in a try/except (this is a must in any Python threading
  application)
- signalling exceptions from worker to main thread in a thread safe
  manner
- terminating the miner if starting the axon fails
- monitoring and reporting on whether the axon still runs

Whether to keep the miner running if axon issues arise later is another
question; the code as-is indicates this is indented behavior: "# In case
of unforeseen errors, the miner will log the error and continue
operations." so this is not changed.

This patch depends on another patch to bittensor that adds .is_running()
and .exception to class axon.